Moved swipe detection to Controller
