+++ core/src/com/spacejunk/game/Controller.java
@@ -5,6 +5,7 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.spacejunk.game.menus.ConsumablesMenu;
 import com.spacejunk.game.menus.OptionsMenu;
+import com.spacejunk.game.utilities.SimpleDirectionGestureDetector;
 
 /**
  * Created by vidxyz on 2/9/18.
@@ -38,6 +39,34 @@ public class Controller {
     public void render(SpriteBatch canvas) {
         optionsMenu.render(canvas);
         consumablesMenu.render(canvas);
+    }
+
+    public void setupSwipeDetection() {
+        Gdx.input.setInputProcessor(new SimpleDirectionGestureDetector(new SimpleDirectionGestureDetector.DirectionListener() {
+
+            @Override
+            public void onUp() {
+                currentGame.getCharacter().moveCharacter(Gdx.graphics.getHeight());
+                Gdx.app.log("swipelog", "onUp caled");
+            }
+
+            @Override
+            public void onRight() {
+
+            }
+
+            @Override
+            public void onLeft() {
+
+            }
+
+            @Override
+            public void onDown() {
+                currentGame.getCharacter().moveCharacter(0);
+                Gdx.app.log("swipelog", "onDown caled");
+            }
+
+        }));
     }
 
     public OptionsMenu getOptionsMenu() {
